Remove ExecutionStrategy abstract class
